Split service into one file per command. Create ServiceCreateOptions following oc conventions.
